requirements.txt
pygame-ce>=2.5.6
numpy>=1.24.0

